Remove unused methods
